tree_operations/requirements.txt
boto3
biopython==1.81
dendropy==4.6.1